JsonIO Finished version
JsonIO allows to read and write json files easily without a lot of configuration.

It does so adding a layer above the jackson library please take into consideration that if you want to have further configuration capabilities, use directly the jackson library.

In this commit there are working and functional the following capabilities:

1. Reading a json file into an specific object
2. Writing a specific object to a json file

There is also some documentation on how to properly use all functionality implemented by this commit