Finishes init panel
Adds a method to switch between panels on entry
Creates a controller for init.

TODO create the new jpanel to reserve a table
TODO update the constant for the identifier of that jpanel

Right now we are talking about GUI, TODOs for logic will come after

IDEA. Image comes from server? Propose for the future